Fix up ipam controller resources
